MockHTTPServer: mock proxy, redirect, file HTTP server
Implement mock HTTP proxy server, HTTP redirect server and file
serving HTTP server.
