+++ test/unittests/c_http_server.h
@@ -5,29 +5,23 @@
 #ifndef TEST_UNITTESTS_C_HTTP_SERVER_H_
 #define TEST_UNITTESTS_C_HTTP_SERVER_H_
 
-
-#include <errno.h>
-#include <fcntl.h>
-#include <netinet/in.h>
 #include <pthread.h>
-#include <signal.h>
-#include <sys/socket.h>
-#include <unistd.h>
 
 #include <cassert>
-#include <cstring>
 #include <string>
 #include <utility>
 #include <vector>
 
 #include "atomic.h"
-#include "duplex_curl.h"
-#include "util/posix.h"
 #include "util/string.h"
 
 typedef std::vector< std::pair<std::string, std::string> > HTTPHeaderList;
 
 struct HTTPRequest {
+  HTTPRequest() {
+    content_length = 0;
+  }
+
   std::string ToString() const {
     std::string result;
     result += method + " " + path + " " + protocol + "\r\n";
@@ -40,6 +34,7 @@ struct HTTPRequest {
     result += body;
     return result;
   }
+
   std::string method;
   std::string path;
   std::string protocol;
@@ -55,6 +50,7 @@ struct HTTPResponse {
     reason = "OK";
     raw = false;
   }
+
   std::string ToString() const {
     if (raw) return body;
     std::string result;
@@ -70,9 +66,11 @@ struct HTTPResponse {
     result += body;
     return result;
   }
+
   void AddHeader(std::string key, std::string value) {
     headers.push_back(std::pair<std::string, std::string>(key, value));
   }
+
   std::string protocol;
   int code;
   std::string reason;
@@ -122,6 +120,7 @@ class HTTPRequestParser {
     kBody,
     kFinished
   };
+
   int state_;
   std::string buffer_;
   std::string headerKeyBuffer_;
@@ -157,147 +156,57 @@ class MockHTTPServer {
   atomic_int32 server_thread_ready_;
   int server_port_;
 
-  void *callback_data_ = NULL;
-  HTTPResponse (*callback_func_)(const HTTPRequest &, void*) = NULL;
+  void *callback_data_;
+  HTTPResponse (*callback_func_)(const HTTPRequest &, void*);
 
   pthread_t server_thread_;
 };
 
 class MockFileServer {
  public:
-  explicit MockFileServer(int port, std::string root_dir) {
-    port_ = port;
-    root_dir_ = root_dir;
-    server_ = new MockHTTPServer(port);
-    server_->SetResponseCallback(FileServerHandler, this);
-    assert(server_->Start());
-  }
-  ~MockFileServer() {
-    delete server_;
-  }
+  explicit MockFileServer(int port, std::string root_dir);
+  ~MockFileServer();
+  int num_processed_requests() { return num_processed_requests_; }
 
  protected:
-  static HTTPResponse FileServerHandler(const HTTPRequest &req, void *data) {
-    MockFileServer *file_server = static_cast<MockFileServer *>(data);
-    HTTPResponse response;
-    if (req.method == "GET") {
-      bool host_header = false;
-      HeaderList::const_iterator it = req.headers.begin();
-      HeaderList::const_iterator itend = req.headers.end();
-      for (; it != itend; ++it) {
-        if (it->first == "Host") {
-          host_header = true;
-          break;
-        }
-      }
-      std::string local_path = file_server->root_dir_;
-      if (host_header) {
-        local_path += "/" + req.path;
-      } else {
-        local_path += "/" + req.path.substr(req.path.find("/") + 1);
-      }
-      if (FileExists(local_path)) {
-        int fd = open(local_path.c_str(), O_RDONLY);
-        assert(fd >= 0);
-        SafeReadToString(fd, &response.body);
-        close(fd);
-      } else {
-        response.code = 404;
-        response.reason = "Not Found";
-      }
-    }
-    return response;
-  }
+  static HTTPResponse FileServerHandler(const HTTPRequest &req, void *data);
+
   std::string root_dir_;
   int port_;
   MockHTTPServer *server_;
+  int num_processed_requests_;
 };
 
 class MockProxyServer {
  public:
-  explicit MockProxyServer(int port) {
-    port_ = port;
-    server_ = new MockHTTPServer(port);
-    server_->SetResponseCallback(ProxyServerHandler, this);
-    assert(server_->Start());
-  }
-  ~MockProxyServer() {
-    delete server_;
-  }
- 
+  explicit MockProxyServer(int port);
+  ~MockProxyServer();
+  int num_processed_requests() { return num_processed_requests_; }
+
  protected:
   static size_t ProxyServerWriteCallback(char *ptr, size_t size, size_t nmemb,
-                                         void* userdata) {
-    std::string *response = static_cast<std::string *>(userdata);
-    (*response) += std::string(ptr, size*nmemb);
-    return size*nmemb;
-  }
-
-  static HTTPResponse ProxyServerHandler(const HTTPRequest &req, void *data) {
-    // MockProxyServer *proxy_server = static_cast<MockProxyServer *>(data);    
-    HTTPResponse response;
-    CURL* handle = curl_easy_init();
-    curl_easy_setopt(handle, CURLOPT_HEADER, 1);
-    curl_easy_setopt(handle, CURLOPT_URL, req.path.c_str());
-    curl_easy_setopt(handle, CURLOPT_CUSTOMREQUEST, req.method.c_str());
-
-    std::string destination_response;
-    curl_easy_setopt(handle, CURLOPT_WRITEFUNCTION, ProxyServerWriteCallback);
-    curl_easy_setopt(handle, CURLOPT_WRITEDATA, &destination_response);
-
-    struct curl_slist *header_list = NULL;
-    HeaderList::const_iterator it = req.headers.begin();
-    HeaderList::const_iterator itend = req.headers.end();
-    for (; it != itend; ++it) {
-      std::string header = it->first + ": " + it->second;
-      header_list = curl_slist_append(header_list, header.c_str());
-    }
-    curl_easy_setopt(handle, CURLOPT_HTTPHEADER, header_list);
-
-    if (req.body.length()) {
-      curl_easy_setopt(handle, CURLOPT_POSTFIELDS, req.body.c_str());
-      curl_easy_setopt(handle, CURLOPT_POSTFIELDSIZE, req.body.length());
-    }
-    curl_easy_perform(handle);
-    curl_easy_cleanup(handle);
-    curl_slist_free_all(header_list);
-    response.raw = true;
-    response.body = destination_response;
-    return response;
-  }
+                                         void* userdata);
+  static HTTPResponse ProxyServerHandler(const HTTPRequest &req, void *data);
 
   int port_;
   MockHTTPServer *server_;
+  int num_processed_requests_;
 };
 
 class MockRedirectServer {
  public:
-  explicit MockRedirectServer(int port, std::string redirect_destination) {
-    port_ = port;
-    server_ = new MockHTTPServer(port);
-    redirect_destination_ = redirect_destination;
-    server_->SetResponseCallback(RedirectServerHandler, this);
-    assert(server_->Start());
-  }
-  ~MockRedirectServer() {
-    delete server_;
-  }
+  explicit MockRedirectServer(int port, std::string redirect_destination);
+  ~MockRedirectServer();
+  int num_processed_requests() { return num_processed_requests_; }
 
  protected:
   static HTTPResponse RedirectServerHandler(const HTTPRequest &req,
-                                            void *data) {
-    MockRedirectServer *redirect_server = static_cast<MockRedirectServer *>(data);    
-    HTTPResponse response;
-    response.code = 301;
-    response.reason = "Moved Permanently";
-    response.AddHeader("Location",
-                       redirect_server->redirect_destination_ + req.path);
-    return response;
-  }
+                                            void *data);
 
   int port_;
   std::string redirect_destination_;
   MockHTTPServer *server_;
+  int num_processed_requests_;
 };
 
 #endif  // TEST_UNITTESTS_C_HTTP_SERVER_H_
